Prime Numbers
Write a Java program to determine whether a given number is prime or not. A prime number is a natural number greater than 1 that is not a product of two smaller natural numbers other than 1 and itself. Implement a program that takes an integer as input and outputs whether it is a prime number or not.